JBRULES-2808 Remove TimeMachine so we have a unified clock structure
git-svn-id: https://svn.jboss.org/repos/labs/labs/jbossrules/trunk@36097 c60d74c8-e8f6-0310-9e8f-d4a2fc68ab70

+++ drools-core/src/main/java/org/drools/time/impl/PseudoClockScheduler.java
@@ -131,7 +131,7 @@ public class PseudoClockScheduler
         return this.timer;
     }
 
-    public synchronized void setStartupTime(int i) {
+    public synchronized void setStartupTime(long i) {
         this.timer = i;
     }
